Move 'getTimeDurationObject' and 'printTimeDuration' into a separate file
